mptcp: consolidate tcp_skb_cb by moving path_mask into a union
The path_mask field is used on the send path while the header field is
used on the received side, so we can reduce the tcp_skb_cb size by
creating union for the path_mask and header fields.

This patch also reverts the size of sk_buff back to 48 since now
the size of tcp_skb_cb is back to 44 bytes.

Signed-off-by: Octavian Purdila <octavian.purdila@intel.com>
Signed-off-by: Christoph Paasch <christoph.paasch@uclouvain.be>