dirty fix to make work the match
The match was no longer working because the nf_mptcp_get_ptr implementation
wasn't working while being compiled in a module. Dunno why yet.
Trying the trampoline solution given here:
http://stackoverflow.com/questions/11463184/how-to-use-exported-symbols-in-the-linux-kernel
makes the kernel oops...
So, this puts the function in the header used by all mptcp related stuff in
netfilter. It's not very pretty but it works atm.